Se modifcó el controlador de area para permitir modificar evaluaciones

+++ app/Http/Controllers/AreaController.php
@@ -427,8 +427,24 @@ class AreaController extends Controller
     }
 
     public function saveFinal_Curso(Request $request,$profesor_id,$curso_id, $catalogoCurso_id){
+        $participante = ParticipantesCurso::where('profesor_id',$profesor_id)->where('curso_id',$curso_id)->get();
+        $evaluacion_id = DB::table('_evaluacion_final_curso')
+            ->select('id')
+            ->where([['participante_curso_id',$participante[0]->id],['curso_id',$curso_id]])
+            ->get();
+
+        if(sizeof($participante) > 0){
+            $evaluacion_id = DB::table('_evaluacion_final_curso')
+                ->select('id')
+                ->where([['participante_curso_id',$participante[0]->id],['curso_id',$curso_id]])
+                ->get();
+            if(sizeof($evaluacion_id) > 0){
+                $eval_fcurso = EvaluacionFinalSeminario::find($evaluacion_id[0]->id);
+                $eval_fcurso->delete();
+            }
+        }
+
         $eval_fcurso = new EvaluacionFinalCurso;
-		$participante = ParticipantesCurso::where('profesor_id',$profesor_id)->where('curso_id',$curso_id)->get();
 		try{
 			$eval_fcurso->participante_curso_id=$participante[0]->id;
 			//Obtenemos la fecha actual para usarla en consultas posteriores	
@@ -573,9 +589,25 @@ class AreaController extends Controller
 			//Horarios Intersemestrales:
 			$eval_fcurso->horarioi = $request->horarioi;
 			$eval_fcurso->curso_id = $curso_id;
+
+            $string_vals = ['mejor','sug','otros','conocimiento','tematica','horarios','horarioi'];
+
+            foreach($eval_fcurso->getAttributes() as $key => $value){
+                if($value == null){
+                    if($key == 'p7'){
+                        $eval_fcurso->$key = -1;
+                    }else if(in_array($key,$string_vals,TRUE)){
+                        $eval_fcurso->$key = '';
+                    }else if($key == 'p8[0]'){
+                        $eval_fcurso->$key = [''];
+                    }else{
+                        $eval_fcurso->$key = 0;
+                    }
+                }
+            }
+
 			$eval_fcurso->save();
 		}catch (\Exception $e){
-
 			//En caso de que no se haya evaluado correctamente el curso regresamos a la vista anterior indicando que la evaluación fue errónea
 			Session::flash('message','Sucedió un error al contestar el formulario. Favor de llenar todas las preguntas o revisar que el usuario en cuestión no lo haya contestado');
 			Session::flash('alert-class', 'alert-danger'); 
@@ -633,27 +665,28 @@ class AreaController extends Controller
 			->where('curso_id',$curso_id)
 			->update(['contesto_hoja_evaluacion' => true]);
 
-	
-        return redirect()->route('area.evaluacion',[$curso_id]);
+	 
+        return redirect()->route('cd.evaluacion',[$curso_id]);
     }
 
     public function saveFinal_Seminario(Request $request,$profesor_id,$curso_id, $catalogoCurso_id){
-        $eval_fseminario = new EvaluacionFinalSeminario;
         $promedio_p1 = new EvaluacionFinalSeminario;
-		
-        $correo = new EvaluacionController(); 
-		$participante = ParticipantesCurso::where('profesor_id',$profesor_id)->where('curso_id',$curso_id)->get();
-		$evaluacion_id = DB::table('_evaluacion_final_seminario')
-            ->select('id')
-            ->where([['participante_curso_id',$participante[0]->id],['curso_id',$curso_id]])
-            ->get();
+        $correo = new EvaluacionFinalSeminario;
 
-        if(sizeof($evaluacion_id) > 0){
-            $eval_fcurso = EvaluacionFinalSeminario::find($evaluacion_id[0]->id);
-            $eval_fcurso->delete();
+		$participante = ParticipantesCurso::where('profesor_id',$profesor_id)->where('curso_id',$curso_id)->get();
+        if(sizeof($participante) > 0){
+            $evaluacion_id = DB::table('_evaluacion_final_seminario')
+                ->select('id')
+                ->where([['participante_curso_id',$participante[0]->id],['curso_id',$curso_id]])
+                ->get();
+            if(sizeof($evaluacion_id) > 0){
+                $eval_fcurso = EvaluacionFinalSeminario::find($evaluacion_id[0]->id);
+                $eval_fcurso->delete();
+            }
         }
+        $eval_fseminario = new EvaluacionFinalSeminario;
 		try{
-			$eval_fseminario->participante_curso_id=$participante[0]->id;
+		  	$eval_fseminario->participante_curso_id=$participante[0]->id;
 			$eval_fseminario->curso_id = $curso_id;
 			
 			
@@ -756,7 +789,26 @@ class AreaController extends Controller
 			$eval_fseminario->horarios = $request->horarios;	
 			//Horarios Intersemestrales:
 			$eval_fseminario->horarioi = $request->horarioi;
+
+            $string_vals = ['mejor','sug','otros','conocimiento','tematica','horarios','horarioi'];
+
+            foreach($eval_fseminario->getAttributes() as $key => $value){
+                if($value == null){
+                    if($key == 'p7'){
+                        $eval_fseminario->$key = -1;
+                    }else if(in_array($key,$string_vals,TRUE)){
+                        $eval_fseminario->$key = '';
+                    }else if($key == 'p8[0]'){
+                        $eval_fseminario->$key = [''];
+                    }else{
+                        $eval_fseminario->$key = 0;
+                    }
+                }
+            }
+
+
 			$eval_fseminario->save();
+
 		} catch(\Exception $e){
 
 			//En caso de que no se haya evaluado correctamente el curso regresamos a la vista anterior indicando que la evaluación fue errónea
@@ -795,45 +847,45 @@ $promedio_p4=[
             $eval_fseminario->p4_9,
             $eval_fseminario->p4_10,
             $eval_fseminario->p4_11];
-        $promedio=[
-            $eval_fseminario->p1_1,
-            $eval_fseminario->p1_2,
-            $eval_fseminario->p1_3,
-            $eval_fseminario->p1_4,
-            $eval_fseminario->p1_5,
-            $eval_fseminario->p2_1,
-            $eval_fseminario->p2_2,
-            $eval_fseminario->p2_3,
-            $eval_fseminario->p2_4,
-            $eval_fseminario->p3_1,
-            $eval_fseminario->p3_2,
-            $eval_fseminario->p3_3,
-            $eval_fseminario->p3_4,
-            $eval_fseminario->p4_1,
-            $eval_fseminario->p4_2,
-            $eval_fseminario->p4_3,
-            $eval_fseminario->p4_4,
-            $eval_fseminario->p4_5,
-            $eval_fseminario->p4_6,
-            $eval_fseminario->p4_7,
-            $eval_fseminario->p4_8,
-            $eval_fseminario->p4_9,
-            $eval_fseminario->p4_10,
-            $eval_fseminario->p4_11
-        ];
+            $promedio=[
+                $eval_fseminario->p1_1,
+                $eval_fseminario->p1_2,
+                $eval_fseminario->p1_3,
+                $eval_fseminario->p1_4,
+                $eval_fseminario->p1_5,
+                $eval_fseminario->p2_1,
+                $eval_fseminario->p2_2,
+                $eval_fseminario->p2_3,
+                $eval_fseminario->p2_4,
+                $eval_fseminario->p3_1,
+                $eval_fseminario->p3_2,
+                $eval_fseminario->p3_3,
+                $eval_fseminario->p3_4,
+                $eval_fseminario->p4_1,
+                $eval_fseminario->p4_2,
+                $eval_fseminario->p4_3,
+                $eval_fseminario->p4_4,
+                $eval_fseminario->p4_5,
+                $eval_fseminario->p4_6,
+                $eval_fseminario->p4_7,
+                $eval_fseminario->p4_8,
+                $eval_fseminario->p4_9,
+                $eval_fseminario->p4_10,
+                $eval_fseminario->p4_11
+            ];
 
         $p1=collect($promedio_p1)->average()*2*10;
         $p2=collect($promedio_p2)->average()*2*10;
         $p3=collect($promedio_p3)->average()*2*10;
         $p4=collect($promedio_p4)->average()*2*10;
         $pg=collect($promedio)->average()*2*10;
-		
+        
           //Actualizar tabla en la bd
         DB::table('participante_curso')
-        	->where('id', $participante[0]->id)
-        	->where('curso_id',$curso_id)
-			->update(['contesto_hoja_evaluacion' => true]);
-		
+            ->where('id', $participante[0]->id)
+            ->where('curso_id',$curso_id)
+		    ->update(['contesto_hoja_evaluacion' => true]);
+
 		//Actualizar campo de hoja de evaluacion
 		DB::table('participante_curso')
 			->where('id', $participante[0]->id)
@@ -841,7 +893,7 @@ $promedio_p4=[
 			->update(['contesto_hoja_evaluacion' => true]);
 
 	
-        return redirect()->route('area.evaluacion',[$curso_id]);
+        return redirect()->route('cd.evaluacion',[$curso_id]);
     }
 
 	public function modificarEvaluacion(Request $request, int $curso_id, int $profesor_id){
@@ -917,20 +969,10 @@ $promedio_p4=[
 
 	public function changeFinal_Curso(Request $request,$profesor_id,$curso_id, $catalogoCurso_id){
         $participante = ParticipantesCurso::where('profesor_id',$profesor_id)->where('curso_id',$curso_id)->get();
-        $evaluacion_id = DB::table('_evaluacion_final_curso')
-            ->select('id')
-            ->where([['participante_curso_id',$participante[0]->id],['curso_id',$curso_id]])
-            ->get();
-
-        $eval_fcurso = EvaluacionFinalCurso::find($evaluacion_id[0]->id);
-
-        $eval_fcurso->delete();
-        
         return $this->saveFinal_Curso($request,$profesor_id,$curso_id, $catalogoCurso_id);
     }
 
 	public function changeFinal_Seminario(Request $request,$profesor_id,$curso_id, $catalogoCurso_id){
-        
         $participante = ParticipantesCurso::where('profesor_id',$profesor_id)->where('curso_id',$curso_id)->get();
         return $this->saveFinal_Seminario($request,$profesor_id,$curso_id, $catalogoCurso_id);
     }
